+++ lib/telegram/bot/types/base.rb
@@ -5,6 +5,10 @@ module Telegram
     module Types
       class Base
         include Virtus.model
+
+        def to_hash(*)
+          super.reject { |_k, v| v.nil? }
+        end
       end
     end
   end